+++ src/main/java/com/tallerwebi/dominio/Restaurante.java
@@ -13,12 +13,12 @@ public class Restaurante {
     private String nombre;
 
     @Column(nullable = false)
-    private Integer estrellas;
+    private Double estrellas;
 
     @Column(nullable = false)
     private String direccion;
 
-    public Restaurante (Long id, String nombre, Integer estrellas, String direccion) {
+    public Restaurante (Long id, String nombre, Double estrellas, String direccion) {
         this.id = id;
         this.nombre = nombre;
         this.estrellas = estrellas;
@@ -28,8 +28,8 @@ public class Restaurante {
     public Long getId() {return id;}
     public String getNombre() {return nombre;}
     public void setNombre(String nombre) {this.nombre = nombre;}
-    public Integer getEstrellas() {return estrellas;}
-    public void setEstrellas(Integer estrellas) {this.estrellas = estrellas;}
+    public Double getEstrellas() {return estrellas;}
+    public void setEstrellas(Double estrellas) {this.estrellas = estrellas;}
     public String getDireccion() {return direccion;}
     public void setDireccion(String direccion) {this.direccion = direccion;}
 }
